backend-temas-coincidentes/pyproject.toml
[project]
name = "backend-temas-coincidentes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "authlib>=1.6.5",
    "fastapi>=0.121.3",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]
